Add files via upload
I wanted to focus on this track because I'm very happy with how the editing worked out on it!  The clip I downloaded had a fast heartbeat that gradually slowed down, but I was looking for the opposite effect.  I attempted to just reverse the clip, but the individual beats sounded off.  So instead, I cut the audio clip into sections of slow heartbeat, medium speed, and fast, ordered them as I liked, then used the envelope tool to increase the speed across the track until eventually fading it out at the end.