Add files via upload
I added this screenshot -- though it's not the most obvious -- to show the increasing dB on the "Typing Track" and the "Hillman Talking" track.  In these two recordings, I first had the recordings fade in; I then used the envelope tool to increase the volume as the tracks went on.  In addition, at the end of the track, to create a more overwhelming and louder experience, I amplified and used distortion on the later ends of the clips.